feat(workspace): Distribute pipeline, not providers
