+++ bin/client/Cargo.toml
@@ -15,7 +15,6 @@ kona-common.workspace = true
 kona-derive.workspace = true
 kona-preimage.workspace = true
 kona-executor.workspace = true
-kona-providers.workspace = true
 kona-common-proc.workspace = true
 
 # Alloy
