+++ bin/client/src/l1/driver.rs
@@ -19,12 +19,14 @@ use kona_derive::{
         AttributesQueue, BatchQueue, BatchStream, ChannelProvider, ChannelReader, FrameQueue,
         L1Retrieval, L1Traversal,
     },
-    traits::{ActivationSignal, BlobProvider, OriginProvider, ResetSignal, Signal, SignalReceiver},
+    traits::{
+        ActivationSignal, BlobProvider, ChainProvider, L2ChainProvider, OriginProvider,
+        ResetSignal, Signal, SignalReceiver,
+    },
 };
 use kona_executor::{KonaHandleRegister, StatelessL2BlockExecutor};
 use kona_mpt::{TrieHinter, TrieProvider};
 use kona_preimage::{CommsClient, PreimageKey, PreimageKeyType};
-use kona_providers::{ChainProvider, L2ChainProvider};
 use op_alloy_consensus::OpTxType;
 use op_alloy_genesis::RollupConfig;
 use op_alloy_protocol::{BlockInfo, L2BlockInfo};
