+++ bin/host/Cargo.toml
@@ -18,15 +18,15 @@ kona-client.workspace = true
 kona-derive.workspace = true
 kona-common.workspace = true
 kona-preimage.workspace = true
-kona-providers-alloy.workspace = true
+kona-derive-alloy.workspace = true
 
 # Alloy
 alloy-rlp.workspace = true
 alloy-eips.workspace = true
 alloy-provider.workspace = true
 alloy-consensus.workspace = true
-alloy-rpc-client.workspace = true 
-alloy-transport-http.workspace = true 
+alloy-rpc-client.workspace = true
+alloy-transport-http.workspace = true
 alloy-rpc-types = { workspace = true, features = ["eth"] }
 alloy-primitives = { workspace = true, features = ["serde"] }
