+++ crates/derive-alloy/README.md
@@ -1,0 +1,5 @@
+## `kona-derive-alloy`
+
+_Notice: Requires an `std` environment._
+
+An online, alloy-backed derivation pipeline, built on [`kona-derive`](../derive).
