+++ crates/derive-alloy/src/blob_provider.rs
