+++ crates/derive-alloy/src/testdata/eth_v1_beacon_sidecars_goerli.json
