+++ crates/derive/Cargo.toml
@@ -34,9 +34,6 @@ async-trait.workspace = true
 unsigned-varint.workspace = true
 alloc-no-stdlib.workspace = true
 
-# Workspace
-kona-providers.workspace = true
-
 # `serde` feature dependencies
 serde = { workspace = true, optional = true, features = ["derive"] }
 
@@ -49,7 +46,7 @@ spin = { workspace = true, optional = true }
 anyhow = { workspace = true, optional = true }
 alloy-rpc-client = { workspace = true, optional = true }
 alloy-node-bindings = { workspace = true, optional = true }
-alloy-transport-http = { workspace = true, optional = true } 
+alloy-transport-http = { workspace = true, optional = true }
 tracing-subscriber = { workspace = true, optional = true, features = ["fmt"] }
 
 [dev-dependencies]
@@ -62,7 +59,6 @@ alloy-node-bindings.workspace = true
 alloy-transport-http.workspace = true
 tokio = { workspace = true, features = ["full"] }
 tracing-subscriber = { workspace = true, features = ["fmt"] }
-kona-providers = { workspace = true, features = ["test-utils"] }
 alloy-primitives = { workspace = true, features = ["rlp", "k256", "map", "arbitrary"] }
 
 [features]
@@ -87,6 +83,5 @@ test-utils = [
   "dep:alloy-node-bindings",
   "dep:tracing-subscriber",
   "dep:alloy-rpc-client",
-  "kona-providers/test-utils",
   "alloy-transport-http/reqwest"
 ]
