+++ crates/derive/src/stages/batch/batch_stream.rs
@@ -3,7 +3,7 @@
 use crate::{
     batch::{Batch, BatchValidity, BatchWithInclusionBlock, SingleBatch, SpanBatch},
     errors::{PipelineEncodingError, PipelineError, PipelineResult},
-    pipeline::L2ChainProvider,
+    traits::L2ChainProvider,
     stages::NextBatchProvider,
     traits::{OriginAdvancer, OriginProvider, Signal, SignalReceiver},
 };
@@ -223,10 +223,9 @@ mod test {
     use super::*;
     use crate::{
         batch::{SingleBatch, SpanBatchElement},
-        test_utils::{CollectingLayer, TestBatchStreamProvider, TraceStorage},
+        test_utils::{CollectingLayer, TestBatchStreamProvider, TestL2ChainProvider, TraceStorage},
         traits::ResetSignal,
     };
-    use kona_providers::test_utils::TestL2ChainProvider;
     use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt};
 
     #[tokio::test]
