--- crates/providers-alloy/README.md
@@ -1,5 +1,0 @@
-## `kona-providers-alloy`
-
-_Notice: Requires an `std` environment._
-
-Alloy providers for `kona-derive`.
