--- crates/providers/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "kona-providers"
-description = "Provider traits for Kona"
-version = "0.0.1"
-edition.workspace = true
-authors.workspace = true
-license.workspace = true
-repository.workspace = true
-homepage.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-# Alloy
-alloy-consensus.workspace = true
-alloy-primitives.workspace = true
-
-# OP Alloy
-op-alloy-genesis.workspace = true
-op-alloy-protocol.workspace = true
-op-alloy-consensus = { workspace = true, features = ["k256"] }
-
-# Misc
-async-trait.workspace = true
-
-# `test-utils` feature
-thiserror = { workspace = true, optional = true }
-
-[features]
-default = []
-test-utils = ["dep:thiserror"]
