--- crates/providers/README.md
@@ -1,3 +1,0 @@
-## `kona-providers`
-
-Provider trait definitions for `kona-derive`.
