--- crates/providers/src/l1_chain_provider.rs
@@ -1,32 +1,0 @@
-//! L1 Chain Provider.
-
-use alloc::{boxed::Box, string::ToString, vec::Vec};
-use alloy_consensus::{Header, Receipt, TxEnvelope};
-use alloy_primitives::B256;
-use async_trait::async_trait;
-use core::fmt::Display;
-use op_alloy_protocol::BlockInfo;
-
-/// Describes the functionality of a data source that can provide information from the blockchain.
-#[async_trait]
-pub trait ChainProvider {
-    /// The error type for the [ChainProvider].
-    type Error: Display + ToString;
-
-    /// Fetch the L1 [Header] for the given [B256] hash.
-    async fn header_by_hash(&mut self, hash: B256) -> Result<Header, Self::Error>;
-
-    /// Returns the block at the given number, or an error if the block does not exist in the data
-    /// source.
-    async fn block_info_by_number(&mut self, number: u64) -> Result<BlockInfo, Self::Error>;
-
-    /// Returns all receipts in the block with the given hash, or an error if the block does not
-    /// exist in the data source.
-    async fn receipts_by_hash(&mut self, hash: B256) -> Result<Vec<Receipt>, Self::Error>;
-
-    /// Returns the [BlockInfo] and list of [TxEnvelope]s from the given block hash.
-    async fn block_info_and_transactions_by_hash(
-        &mut self,
-        hash: B256,
-    ) -> Result<(BlockInfo, Vec<TxEnvelope>), Self::Error>;
-}
